+++ TraceFencePlugins/QuotaMonitor/Bundle/QuotaMonitorPluginBundleEntrypoint.swift
@@ -1,0 +1,3 @@
+import QuotaMonitorPlugin
+
+private let quotaMonitorPluginFactoryAnchor: Any.Type = QuotaMonitorPluginFactory.self